Move remaining emoji picker classes
