+++ app/src/fdroid/java/com/geeksville/mesh/ui/map/components/EditWaypointDialog.kt
@@ -64,10 +64,10 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.geeksville.mesh.MeshProtos.Waypoint
 import com.geeksville.mesh.copy
-import com.geeksville.mesh.ui.common.components.EmojiPickerDialog
 import com.geeksville.mesh.waypoint
 import org.meshtastic.core.strings.R
 import org.meshtastic.core.ui.component.EditTextPreference
+import org.meshtastic.core.ui.emoji.EmojiPickerDialog
 import org.meshtastic.core.ui.theme.AppTheme
 import java.text.SimpleDateFormat
 import java.util.Calendar
