+++ app/src/google/java/com/geeksville/mesh/ui/map/components/EditWaypointDialog.kt
@@ -64,8 +64,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.geeksville.mesh.MeshProtos.Waypoint
 import com.geeksville.mesh.copy
-import com.geeksville.mesh.ui.common.components.EmojiPickerDialog
 import org.meshtastic.core.strings.R
+import org.meshtastic.core.ui.emoji.EmojiPickerDialog
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Locale
