+++ core/ui/detekt-baseline.xml
@@ -3,7 +3,6 @@
   <ManuallySuppressedIssues/>
   <CurrentIssues>
     <ID>ComposableParamOrder:AlertDialogs.kt$SimpleAlertDialog</ID>
-    <ID>ComposableParamOrder:BatteryInfo.kt$BatteryInfo</ID>
     <ID>ComposableParamOrder:EditTextPreference.kt$EditTextPreference</ID>
     <ID>ComposableParamOrder:MaterialBatteryInfo.kt$MaterialBatteryInfo</ID>
     <ID>ComposableParamOrder:NodeChip.kt$NodeChip</ID>
@@ -12,22 +11,14 @@
     <ID>ComposableParamOrder:SwitchPreference.kt$SwitchPreference</ID>
     <ID>ContentSlotReused:AdaptiveTwoPane.kt$second</ID>
     <ID>LongMethod:EditTextPreference.kt$@Composable fun EditTextPreference( title: String, value: String, enabled: Boolean, isError: Boolean, keyboardOptions: KeyboardOptions, keyboardActions: KeyboardActions, onValueChanged: (String) -&gt; Unit, modifier: Modifier = Modifier, summary: String? = null, maxSize: Int = 0, // max_size - 1 (in bytes) onFocusChanged: (FocusState) -&gt; Unit = {}, trailingIcon: (@Composable () -&gt; Unit)? = null, visualTransformation: VisualTransformation = VisualTransformation.None, )</ID>
-    <ID>MagicNumber:BatteryInfo.kt$100</ID>
-    <ID>MagicNumber:BatteryInfo.kt$101</ID>
-    <ID>MagicNumber:BatteryInfo.kt$14</ID>
-    <ID>MagicNumber:BatteryInfo.kt$15</ID>
-    <ID>MagicNumber:BatteryInfo.kt$34</ID>
-    <ID>MagicNumber:BatteryInfo.kt$35</ID>
-    <ID>MagicNumber:BatteryInfo.kt$4</ID>
-    <ID>MagicNumber:BatteryInfo.kt$5</ID>
-    <ID>MagicNumber:BatteryInfo.kt$79</ID>
-    <ID>MagicNumber:BatteryInfo.kt$80</ID>
     <ID>MagicNumber:EditIPv4Preference.kt$0xff</ID>
     <ID>MagicNumber:EditIPv4Preference.kt$16</ID>
     <ID>MagicNumber:EditIPv4Preference.kt$24</ID>
     <ID>MagicNumber:EditIPv4Preference.kt$8</ID>
     <ID>MagicNumber:LazyColumnDragAndDropDemo.kt$50</ID>
     <ID>ModifierMissing:AdaptiveTwoPane.kt$AdaptiveTwoPane</ID>
+    <ID>ModifierMissing:EmojiPicker.kt$EmojiPicker</ID>
+    <ID>ModifierMissing:EmojiPicker.kt$EmojiPickerDialog</ID>
     <ID>ModifierMissing:IndoorAirQuality.kt$IndoorAirQuality</ID>
     <ID>ModifierMissing:LoraSignalIndicator.kt$LoraSignalIndicator</ID>
     <ID>ModifierMissing:LoraSignalIndicator.kt$Rssi</ID>
@@ -38,7 +29,6 @@
     <ID>ModifierMissing:SettingsItem.kt$SettingsItemSwitch</ID>
     <ID>ModifierMissing:SimpleAlertDialog.kt$SimpleAlertDialog</ID>
     <ID>ModifierMissing:SlidingSelector.kt$OptionLabel</ID>
-    <ID>ModifierNotUsedAtRoot:SignalInfo.kt$modifier = modifier</ID>
     <ID>ModifierNotUsedAtRoot:TextDividerPreference.kt$modifier = modifier.fillMaxWidth().padding(all = 16.dp)</ID>
     <ID>ModifierNotUsedAtRoot:TextDividerPreference.kt$modifier = modifier.fillMaxWidth().wrapContentWidth(Alignment.End)</ID>
     <ID>ModifierReused:PreferenceCategory.kt$Card(modifier = modifier.padding(bottom = 8.dp)) { Column( modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp, vertical = 16.dp), horizontalAlignment = Alignment.CenterHorizontally, ) { ProvideTextStyle(MaterialTheme.typography.bodyLarge) { content() } } }</ID>
@@ -54,8 +44,6 @@
     <ID>ParameterNaming:PreferenceFooter.kt$onNegativeClicked</ID>
     <ID>ParameterNaming:PreferenceFooter.kt$onPositiveClicked</ID>
     <ID>ParameterNaming:SlidingSelector.kt$onOptionSelected</ID>
-    <ID>PreviewPublic:BatteryInfo.kt$BatteryInfoPreview</ID>
-    <ID>PreviewPublic:BatteryInfo.kt$BatteryInfoPreviewSimple</ID>
     <ID>PreviewPublic:IndoorAirQuality.kt$IAQScalePreview</ID>
     <ID>PreviewPublic:LazyColumnDragAndDropDemo.kt$LazyColumnDragAndDropDemo</ID>
     <ID>PreviewPublic:MaterialBatteryInfo.kt$MaterialBatteryInfoPreview</ID>
