+++ core/ui/src/main/kotlin/org/meshtastic/core/ui/emoji/EmojiPicker.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.ui.common.components
+package org.meshtastic.core.ui.emoji
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.background
@@ -29,9 +29,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.emoji2.emojipicker.RecentEmojiProviderAdapter
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
-import com.geeksville.mesh.util.CustomRecentEmojiProvider
 import org.meshtastic.core.ui.component.BottomSheetDialog
-import org.meshtastic.core.ui.emoji.EmojiPickerViewModel
 
 @Composable
 fun EmojiPicker(
